a bit of tidying, plus implemented the 'mixup dataloader' that handles multi-labelling
