+++ braindecode/augment/__init__.py
@@ -2,5 +2,4 @@ from .transforms.masking_along_axis import mask_along_time, mask_along_frequency
 from .transforms.merge_two_signals import merge_two_signals, MERGE_TWO_SIGNALS_REQUIRED_VARIABLES
 from .transforms.mixup_beta import mixup_beta, MIXUP_BETA_REQUIRED_VARIABLES
 # TODO: sort imports
-from .mixup_loss import general_mixup_criterion
-from .base import Transform
+from .base import Transform, AugmentedDataset, general_mixup_criterion, mixup_iterator
