+++ braindecode/augment/base.py
@@ -5,6 +5,7 @@
 import torch
 from functools import partial
 import numpy as np
+from torch.utils.data.dataset import Dataset
 
 
 class Transform:
@@ -67,18 +68,82 @@ class Transform:
         return self.operation(datum)
 
 
-"""Custom iterator for training epochs
-"""
+class AugmentedDataset(Dataset):
+    """An overlay (not a subclass though) to apply on a basic WindowsDataset
+    or BaseConcatDataset or a Subset of one of these two classes. Follows the
+    canonical way to augment data of Pytorch (see [this article]
+    (https://pytorch.org/tutorials/beginner/data_loading_tutorial.html) for
+    more informations.) Main point is, if you pass a list of `n` transforms and
+    a dataset of size `s`, you will obtain a dataset of size `n x s`. Note that
+    transforms are reapplied every time the `__getitem__` function is called,
+    so if you have some randomness in it, you will obtain different transformed
+    data at every epoch. Note that you can compose transforms using the Compose
+    class of torchvision.
 
-# Authors: Simon Brandt <simonbrandt@protonmail.com>
-#
-# License: BSD (3-clause)
+    Parameters
+    ----------
+    ds : WindowsDataset or BaseConcatDataset or a Subset of one of those.
+        The unaugmented dataset
+    list_of_transforms :
+        list(Transform or torchvision.transforms.Compose(Transform)),
+        optional. A list of transforms that will be applied to data, by
+        default identity (default=[Transform(lambda datum:datum)]).
+    """
+
+    def __init__(self, ds,
+                 list_of_transforms=[
+                     Transform(lambda datum:datum)]) -> None:
+        # Initialize the augmented dataset, and computes required
+        # variables for applying transforms
+        self.list_of_transforms = list_of_transforms
+        self.ds = ds
+        self.required_variables = {}
+        self.list_of_labels = \
+            list(set([elem[1] for elem in iter(self.ds)])).sort()
+        self.__initialize_required_variables()
+
+    def __len__(self):
+        return(len(self.ds) * len(self.list_of_transforms))
+
+    def __getitem__(self, index):
+        # First get the indexes of the transform and of the unaugmented data
+        tf_index = index % len(self.list_of_transforms)
+        img_index = index // len(self.list_of_transforms)
+
+        # Get the unaugmented data
+        X, y, crops_ind = self.ds[img_index]
+        y = tuple([int(self.list_of_labels[i] == y)
+                   for i in self.list_of_labels])
+
+        class Datum:
+            def __init__(self, X, y, crops_ind, ds, required_variables):
+                self.X = X
+                self.y = y
+                self.crops_ind = crops_ind
+                self.ds = ds
+                self.required_variables = required_variables
+        # Initialize Datum object with metadata required to compute transforms,
+        # then compute the transform.
+        transf_datum = self.list_of_transforms[tf_index](
+            Datum(X, y, crops_ind, self.ds, self.required_variables))
+        # Returns augmented data.
+        X, y, crops_ind = transf_datum.X, transf_datum.y, transf_datum.crops_ind
+
+        return X, y, crops_ind
+
+    def __initialize_required_variables(self):
+
+        for transform in self.list_of_transforms:
+            for key in transform.required_variables.keys():
+                self.required_variables[key] = \
+                    transform.required_variables[key](
+                        self.ds, self.list_of_transforms)
 
 
 class mixup_iterator(torch.utils.data.DataLoader):
     """Implements Iterator for Mixup for EEG data. See [mixup].
     Code adapted from
-    #TODO ref sbbrandt
+    #TODO ref sbbrandt + rewrite docstring
     Parameters
     ----------
     dataset: Dataset
@@ -105,3 +170,23 @@ class mixup_iterator(torch.utils.data.DataLoader):
         crop_inds = torch.tensor(crop_inds).type(torch.int64)
 
         return x, y, crop_inds
+
+
+class general_mixup_criterion:
+
+    def __init__(self, loss=torch.nn.functional.nll_loss):
+        self.loss = loss
+
+    def __call__(self, preds, target):
+        return self.loss_function(preds, target)
+
+    def loss_function(self, preds, target):
+        ret = None
+        for label in target.keys():
+            prop = target[label]
+            loss_val = self.loss(preds, label, reduction='none')
+            if ret is None:
+                ret = torch.mul(prop, loss_val)
+            else:
+                ret += torch.mul(prop, loss_val)
+        return ret.mean()
