+++ braindecode/augment/transforms/merge_two_signals.py
@@ -12,10 +12,11 @@ def merge_two_signals(datum, params):
     train_sample = datum.ds
     label_index_dict = datum.required_variables["label_index_dict"]
     other_signal = np.zeros(datum.X.shape)
-    for label in y.keys():
+    for label_idx in range(len(y)):
         # y[label] is the proportion of the label in y
-        other_signal_index = np.random.choice(label_index_dict[label])
-        other_signal += y[label] * train_sample[other_signal_index][0]
+        other_signal_index = np.random.choice(
+            label_index_dict[train_sample.list_of_labels[label_idx]])
+        other_signal += y[label_idx] * train_sample[other_signal_index][0]
 
     datum.X = (1 - params["magnitude"]) * \
         signal + params["magnitude"] * other_signal
