+++ braindecode/datasets/__init__.py
@@ -1,7 +1,7 @@
 """
 Loader code for some datasets.
 """
-from .base import WindowsDataset, BaseDataset, BaseConcatDataset, AugmentedDataset
+from .base import WindowsDataset, BaseDataset, BaseConcatDataset
 from .moabb import MOABBDataset
 from .tuh import TUHAbnormal
 from .sleep_physionet import SleepPhysionet
